[audio] Avoid taking over phone audio path option
Normally the Android auto head unit would take over the phone's audio.
In many cases this is undesirable because the existing bluetooth
connection in the phone with the car stereo might do a better job of
handling audio/media control. This patch includes the option to make
the head unit to be as close to a screen-only device as possible.

When the option is enabled,
OpenAuto will not take over media and navigation voice channel.